# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: upcontrol API
  version: 1.0.0
  description: |
    The front-facing contract for upcontrol. Generated into a strict Go server
    (oapi-codegen) and into TS types for `front/src/lib/api.d.ts`. Response
    shapes are pinned to `front/src/lib/mockData.ts` — a contract test
    (build-plan §2.7) checks each against golden JSON taken from the mocks.

    Rules from the plan (§4.4):
      - Version in the path (/v1), not a header.
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
      - Errors are one shape: `{error: {code, message, upgrade?}}`. The optional
        text reason. No `disabled`, no `navigate('/plan')`.
      - Pagination only on unbounded lists: incidents, logs. Monitors, channels
        and recipients are plan-bounded and returned whole.
      - `null` is never used for "no value": the field is omitted. "Zero is
        silence" — an absent field does not render.
      - All timestamps are RFC 3339 UTC.
      - `POST /i` and `POST /hooks/{provider}` are `x-hand-written: true`: they
        exist here for the /docs/api page but are implemented by hand (the
        generator's validation is forbidden on /i by §4.5).
servers:
  - url: http://localhost:5173
    description: local dev (front dev server proxies to ucapi)

# ---------------------------------------------------------------------------
# Common building blocks
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uc_session
    apiKey:
      type: apiKey
      in: header
      name: X-Upcontrol-Key
  responses:
...
    Unauthorized:
      description: No session, or the key is not ours.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PaymentRequired:
      description: |
        A plan gate blocked the action. The `upgrade` object carries a text
        reason; the client routes this to its upgrade prompt. No other
        paid-wall path exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "plan_limit_exceeded" }
            message: { type: string }
            upgrade:
              type: object
              description: Present only on 402. Drives the client's upgrade prompt.
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Free allows 3 HTTP checks."
                highlight:
                  type: string
                  description: Optional row to highlight on the pricing table.

    # --- account ------------------------------------------------------------
...
    RecipientRole:
      type: string
      enum: [notify, login]
...
    TelegramInviteCreated:
      type: object
      required: [id, role, link, expiresAt]
      properties:
        id: { type: string }
        role: { $ref: "#/components/schemas/RecipientRole" }
        link:
          type: string
          description: >
            `t.me/<bot>?start=inv_<token>` — the one-time deep link. The token
            appears exactly here and nowhere else: not in logs, not in telemetry,
            not retrievable after this response.
          example: "https://t.me/upcontrol_bot?start=inv_9f8e7d6c5b4a"
        expiresAt: { type: string, format: date-time }
...
paths:
...
  /v1/telegram/invites:
    post:
      tags: [alerts]
      summary: >
        Mint a one-time Telegram invite link (role chosen at invite, default
        notify). Counts active + pending recipients against the plan's
        telegram_recipients axis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: "#/components/schemas/RecipientRole" }
      responses:
        "201":
          description: Created — the link carries the token exactly once.
          content:
            application/json:
              { schema: { $ref: "#/components/schemas/TelegramInviteCreated" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
...